ADDED: Task lifecycle functions

+++ src/stdlib.h
@@ -28,12 +28,11 @@
 #include <cstring>
 #define my_memset(dst, src, n) printf("%p, %p, %d", dst, src, n)
 #define my_memcpy(s, c, n) printf("%p, %d, %u", s, c, n)
-
 #else
 
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
-typedef unsigned long uint32_t;
+typedef unsigned int uint32_t;
 typedef unsigned long long uint64_t;
 
 typedef signed char int8_t;
